Ironsworn: move notes into a tab
