+++ src/module/vue/character-sheet.vue
@@ -42,82 +42,10 @@
           <attr-box :actor="actor" attr="wits"></attr-box>
         </div>
 
-        <div class="flexrow">
-          <div class="flexcol">
-            <section class="sheet-area flexcol">
-              <!-- Bonds -->
-              <bonds :actor="actor"></bonds>
-
-              <hr class="nogrow" />
-              <!-- Assets -->
-              <div
-                class="flexcol ironsworn__drop__target"
-                data-drop-type="asset"
-              >
-                <h4 class="nogrow">{{ $t('IRONSWORN.Assets') }}</h4>
-
-                <transition-group name="slide" tag="div" class="nogrow">
-                  <div
-                    class="flexrow"
-                    v-for="(asset, i) in assets"
-                    :key="asset._id"
-                  >
-                    <order-buttons
-                      v-if="editMode"
-                      :i="i"
-                      :length="assets.length"
-                      @sortUp="assetSortUp"
-                      @sortDown="assetSortDown"
-                    />
-                    <asset :actor="actor" :asset="asset" />
-                  </div>
-                </transition-group>
-                <div class="flexcol nogrow" style="text-align: center">
-                  <btn-compendium
-                    class="block nogrow"
-                    compendium="ironswornassets"
-                  >
-                    {{ $t('IRONSWORN.Assets') }}
-                  </btn-compendium>
-                </div>
-              </div>
-            </section>
-          </div>
-          <div class="flexcol">
-            <!-- Vows & Progress -->
-            <div
-              class="flexcol sheet-area ironsworn__drop__target"
-              data-drop-type="progress"
-            >
-              <transition-group name="slide" tag="div" class="nogrow">
-                <div
-                  class="flexrow nogrow"
-                  v-for="(item, i) in progressItems"
-                  :key="item._id"
-                >
-                  <order-buttons
-                    v-if="editMode"
-                    :i="i"
-                    :length="progressItems.length"
-                    @sortUp="progressSortUp"
-                    @sortDown="progressSortDown"
-                  />
-                  <progress-box :item="item" :actor="actor" />
-                </div>
-              </transition-group>
-
-              <progress-controls :actor="actor" />
-            </div>
-
-            <hr class="nogrow" />
-            <h4 class="nogrow">{{ $t('IRONSWORN.Notes') }}</h4>
-            <mce-editor
-              v-model="actor.data.biography"
-              @save="saveNotes"
-              @change="throttledSaveNotes"
-            />
-          </div>
-        </div>
+        <tabs style="margin-top: 0.5rem">
+          <tab :title="$t('IRONSWORN.Character')"><ironsworn-main /></tab>
+          <tab :title="$t('IRONSWORN.Notes')"><ironsworn-notes /></tab>
+        </tabs>
 
         <!-- Conditions & Banes & Burdens -->
         <section class="sheet-area nogrow">
@@ -173,39 +101,24 @@
 .slide-leave-active {
   max-height: 83px;
 }
-
-textarea.notes {
-  border-color: rgba(0, 0, 0, 0.1);
-  border-radius: 1px;
-  font-family: var(--font-primary);
-  resize: none;
-  flex: 1;
-  min-height: 150px;
-}
 </style>
 
 <script setup lang="ts">
 import { $ActorKey } from './provisions'
-;('vue')
 import AttrBox from './components/attr-box.vue'
 import BtnMomentumburn from './components/buttons/btn-momentumburn.vue'
 import Stack from './components/stack/stack.vue'
 import btnRollstat from './components/buttons/btn-rollstat.vue'
-import btnIsicon from './components/buttons/btn-isicon.vue'
 import { IronswornActor } from '../actor/actor'
 import { provide, computed, inject } from 'vue'
 import { RollDialog } from '../helpers/rolldialog'
 import CharacterHeader from './components/character-header.vue'
-import XpBox from './components/xp-box.vue'
 import Conditions from './components/conditions/conditions.vue'
 import { throttle } from 'lodash'
-import MceEditor from './components/mce-editor.vue'
-import ProgressControls from './components/progress-controls.vue'
-import ProgressBox from './components/progress/progress-box.vue'
-import BtnCompendium from './components/buttons/btn-compendium.vue'
-import Asset from './components/asset/asset.vue'
-import OrderButtons from './components/order-buttons.vue'
-import Bonds from './components/bonds.vue'
+import Tabs from './components/tabs/tabs.vue'
+import Tab from './components/tabs/tab.vue'
+import IronswornMain from './components/character-sheet-tabs/ironsworn-main.vue'
+import IronswornNotes from './components/character-sheet-tabs/ironsworn-notes.vue'
 
 const props = defineProps<{
   actor: ReturnType<typeof IronswornActor.prototype.toObject>
@@ -218,24 +131,6 @@ provide(
 
 const $actor = inject($ActorKey)
 
-const progressItems = computed(() => {
-  return props.actor.items
-    .filter((x) => x.type === 'progress')
-    .sort((a, b) => (a.sort || 0) - (b.sort || 0))
-})
-const assets = computed(() => {
-  return props.actor.items
-    .filter((x) => x.type === 'asset')
-    .sort((a, b) => (a.sort || 0) - (b.sort || 0))
-})
-const editMode = computed(() => {
-  return props.actor.flags['foundry-ironsworn']?.['edit-mode']
-})
-
-//   'actor.data.biography'() {
-//   this.saveNotes()
-// },
-
 function burnMomentum() {
   $actor?.burnMomentum()
 }
@@ -245,38 +140,5 @@ function rollStat(stat) {
 function openCompendium(name) {
   const pack = game.packs?.get(`foundry-ironsworn.${name}`)
   pack?.render(true)
-}
-
-function saveNotes() {
-  $actor?.update({ 'data.biography': props.actor.data.biography })
-}
-const throttledSaveNotes = throttle(saveNotes, 1000)
-
-async function applySort(oldI, newI, sortBefore, collection) {
-  const sorted = collection.sort(
-    (a, b) => (a.data.sort || 0) - (b.data.sort || 0)
-  )
-  const updates = SortingHelpers.performIntegerSort(sorted[oldI], {
-    target: sorted[newI],
-    siblings: sorted,
-    sortBefore,
-  })
-  await Promise.all(updates.map(({ target, update }) => target.update(update)))
-}
-function assetSortUp(i) {
-  const items = $actor?.items.filter((x) => x.type === 'asset')
-  applySort(i, i - 1, true, items)
-}
-function assetSortDown(i) {
-  const items = $actor?.items.filter((x) => x.type === 'asset')
-  applySort(i, i + 1, false, items)
-}
-function progressSortUp(i) {
-  const items = $actor?.items.filter((x) => x.type === 'progress')
-  applySort(i, i - 1, true, items)
-}
-function progressSortDown(i) {
-  const items = $actor?.items.filter((x) => x.type === 'progress')
-  applySort(i, i + 1, false, items)
 }
 </script>
