+++ src/module/vue/components/character-sheet-tabs/ironsworn-notes.vue
@@ -1,0 +1,24 @@
+<template>
+  <div class="flexcol">
+    <MceEditor
+      v-model="actor.data.biography"
+      @change="debouncedSave"
+      @save="immediateSave"
+    />
+  </div>
+</template>
+
+<script setup lang="ts">
+import { debounce } from 'lodash'
+import { inject, Ref } from 'vue'
+import { $ActorKey } from '../../provisions'
+import MceEditor from '../mce-editor.vue'
+
+const actor = inject('actor') as Ref
+const $actor = inject($ActorKey)
+
+const immediateSave = () => {
+  $actor?.update({ 'data.biography': actor.value.data.biography })
+}
+const debouncedSave = debounce(immediateSave, 1000)
+</script>
